Fix the installation error and rename the package.
The package in now renamed to papergit. The installation error is now fixed and
was because of a bad __init__.py file with three trailing underscores.
