+++ papergit/config/__init__.py
@@ -1,6 +1,6 @@
 """Configuration package."""
 
-from paper_to_git.config.config import BaseConfig
+from papergit.config.config import BaseConfig
 
 __all__ = [
     'config'
